backend/requirements.txt

# FastAPI app dependencies
fastapi>=0.95.0
uvicorn[standard]>=0.18.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.0

# Database
SQLAlchemy>=1.4.0
pymysql>=1.0.2

# Optional: read .env
python-dotenv>=1.0.0

